[system] Remove visuallyHidden and move to /utils
