+++ packages/material-ui-system/src/index.js
@@ -20,6 +20,5 @@ export * from './spacing';
 export { default as style } from './style';
 export { default as typography } from './typography';
 export * from './typography';
-export { default as visuallyHidden } from './visuallyHidden';
 export { default as unstable_styleFunctionSx } from './styleFunctionSx';
 export { default as unstable_getThemeValue } from './getThemeValue';
