+++ packages/material-ui-utils/src/index.ts
@@ -30,3 +30,4 @@ export {
   getNormalizedScrollLeft as unstable_getNormalizedScrollLeft,
 } from './scrollLeft';
 export { default as usePreviousProps } from './usePreviousProps';
+export { default as visuallyHidden } from './visuallyHidden';
